refactor: Decompose 2302-line pane_state.rs into focused modules
Break down the monolithic pane_state.rs file into 8 single-responsibility modules:

- capabilities.rs (55 lines) - Capability checking and mode management
- cursor_basic.rs (255 lines) - Basic directional cursor movement
- cursor_line.rs (253 lines) - Line-based cursor operations
- display.rs (321 lines) - Display cache management and line wrapping
- scrolling.rs (344 lines) - Scrolling operations and cursor visibility
- text_operations.rs (591 lines) - Text editing operations
- visual_selection.rs (142 lines) - Visual selection management
- word_navigation.rs (99 lines) - Word navigation functionality
- mod.rs (133 lines) - Main module structure and PaneState definition

This decomposition improves maintainability by following the Single
Responsibility Principle and makes the codebase more navigable for developers.
All functionality is preserved with no behavior changes.
